Moved ClassLoading and Reflections Helper functions to common
