+++ pulsar-common/src/main/java/org/apache/pulsar/common/util/Reflections.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.functions.utils;
+package org.apache.pulsar.common.util;
 
 import java.io.IOException;
 import java.lang.reflect.Array;
@@ -165,7 +165,7 @@ public class Reflections {
 
     public static Object createInstance(String userClassName, java.io.File jar) {
         try {
-            return createInstance(userClassName, FunctionCommon.loadJar(jar));
+            return createInstance(userClassName, ClassLoaderUtils.loadJar(jar));
         } catch (Exception ex) {
             return null;
         }
@@ -180,7 +180,7 @@ public class Reflections {
      */
     public static boolean classExistsInJar(java.io.File jar, String fqcn) {
         try {
-            java.net.URLClassLoader loader = (URLClassLoader) FunctionCommon.loadJar(jar);
+            java.net.URLClassLoader loader = (URLClassLoader) ClassLoaderUtils.loadJar(jar);
             Class.forName(fqcn, false, loader);
             loader.close();
             return true;
@@ -199,7 +199,7 @@ public class Reflections {
         try {
             Class.forName(fqcn);
             return true;
-        } catch( ClassNotFoundException e ) {
+        } catch (ClassNotFoundException e) {
            return false;
         }
     }
@@ -214,7 +214,7 @@ public class Reflections {
     public static boolean classInJarImplementsIface(java.io.File jar, String fqcn, Class xface) {
         boolean ret = false;
         try {
-            java.net.URLClassLoader loader = (URLClassLoader) FunctionCommon.loadJar(jar);
+            java.net.URLClassLoader loader = (URLClassLoader) ClassLoaderUtils.loadJar(jar);
             if (xface.isAssignableFrom(Class.forName(fqcn, false, loader))){
                 ret = true;
             }
@@ -281,7 +281,7 @@ public class Reflections {
                 throw new ClassNotFoundException(className);
             }
         } else if (isPrimitive(className)) {
-            return (Class)PRIMITIVE_NAME_TYPE_MAP.get(className);
+            return (Class) PRIMITIVE_NAME_TYPE_MAP.get(className);
         } else if (className.charAt(0) == 'L' && className.charAt(className.length() - 1) == ';') {
             return classLoader.loadClass(className.substring(1, className.length() - 1));
         } else {
@@ -291,10 +291,12 @@ public class Reflections {
                 if (className.charAt(0) != '[') {
                     throw var4;
                 } else {
+                    // CHECKSTYLE.OFF: EmptyStatement
                     int arrayDimension;
-                    for(arrayDimension = 0; className.charAt(arrayDimension) == '['; ++arrayDimension) {
+                    for (arrayDimension = 0; className.charAt(arrayDimension) == '['; ++arrayDimension) {
                         ;
                     }
+                    // CHECKSTYLE.ON: EmptyStatement
 
                     Class componentType = loadClass(className.substring(arrayDimension), classLoader);
                     return Array.newInstance(componentType, new int[arrayDimension]).getClass();
