+++ pulsar-functions/utils/src/main/java/org/apache/pulsar/functions/utils/FunctionCommon.java
@@ -31,7 +31,6 @@ import java.net.*;
 import java.nio.channels.Channels;
 import java.nio.channels.ReadableByteChannel;
 import java.nio.file.Path;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.UUID;
@@ -42,6 +41,7 @@ import org.apache.pulsar.client.impl.TopicMessageIdImpl;
 import org.apache.pulsar.common.functions.FunctionConfig;
 import org.apache.pulsar.common.functions.Utils;
 import org.apache.pulsar.common.nar.NarClassLoader;
+import org.apache.pulsar.common.util.ClassLoaderUtils;
 import org.apache.pulsar.functions.api.Function;
 import org.apache.pulsar.functions.api.WindowFunction;
 import org.apache.pulsar.functions.proto.Function.FunctionDetails.Runtime;
@@ -205,16 +205,6 @@ public class FunctionCommon {
         return typeArg;
     }
 
-    public static ClassLoader extractClassLoader(Path archivePath, File packageFile) throws Exception {
-        if (archivePath != null) {
-            return loadJar(archivePath.toFile());
-        }
-        if (packageFile != null) {
-            return loadJar(packageFile);
-        }
-        return null;
-    }
-
     public static void downloadFromHttpUrl(String destPkgUrl, File targetFile) throws IOException {
         URL website = new URL(destPkgUrl);
 
@@ -226,22 +216,10 @@ public class FunctionCommon {
         log.info("Downloading function package from {} to {} completed!", destPkgUrl, targetFile.getAbsoluteFile());
     }
 
-    /**
-     * Load a jar.
-     *
-     * @param jar file of jar
-     * @return classloader
-     * @throws MalformedURLException
-     */
-    public static ClassLoader loadJar(File jar) throws MalformedURLException {
-        java.net.URL url = jar.toURI().toURL();
-        return new URLClassLoader(new URL[]{url});
-    }
-
     public static ClassLoader extractClassLoader(String destPkgUrl) throws IOException, URISyntaxException {
         File file = extractFileFromPkgURL(destPkgUrl);
         try {
-            return loadJar(file);
+            return ClassLoaderUtils.loadJar(file);
         } catch (MalformedURLException e) {
             throw new IllegalArgumentException(
                     "Corrupt User PackageFile " + file + " with error " + e.getMessage());
@@ -268,34 +246,6 @@ public class FunctionCommon {
         } else {
             throw new IllegalArgumentException("Unsupported url protocol "+ destPkgUrl +", supported url protocols: [file/http/https]");
         }
-    }
-
-    public static void implementsClass(String className, Class<?> klass, ClassLoader classLoader) {
-        Class<?> objectClass;
-        try {
-            objectClass = loadClass(className, classLoader);
-        } catch (ClassNotFoundException | NoClassDefFoundError e) {
-            throw new IllegalArgumentException("Cannot find/load class " + className);
-        }
-
-        if (!klass.isAssignableFrom(objectClass)) {
-            throw new IllegalArgumentException(
-                    String.format("%s does not implement %s", className, klass.getName()));
-        }
-    }
-
-    public static Class<?> loadClass(String className, ClassLoader classLoader) throws ClassNotFoundException {
-        Class<?> objectClass;
-        try {
-            objectClass = Class.forName(className);
-        } catch (ClassNotFoundException | NoClassDefFoundError e) {
-            if (classLoader != null) {
-                objectClass = classLoader.loadClass(className);
-            } else {
-                throw e;
-            }
-        }
-        return objectClass;
     }
 
     public static NarClassLoader extractNarClassLoader(Path archivePath, File packageFile,
